api/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
meilisearch==0.31.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
jinja2==3.1.6
google-genai==0.1.0
